change Time from i64 to f64 (#145)
Closes #144

+++ src/tsdef.rs
@@ -1,7 +1,7 @@
 /// Integer type used to refer to a genomic coordinate/position
 pub type Position = i64;
-/// Time is recorded as a discrete, signed integer.
-pub type Time = i64;
+/// Time is recorded as [`f64']
+pub type Time = f64;
 /// Integer type used to refer to [Node](struct.Node.html) objects.
 pub type IdType = i32;
 pub type IdVec = Vec<IdType>;
